requirements.txt
pycryptodome==3.17
pypiwin32==223
requests==2.28.2